asppi-backend/pyproject.toml
[tool.poetry]
name = "asppi-backend"
version = "0.1.0"
description = ""
authors = ["Ivan Nikolaev-Axenov <nikolaevaxenov@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.105.0"
uvicorn = "^0.24.0.post1"
fdb = "^2.0.2"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
